Makes use of Data.Void from `void` / `base`.
The `void` package takes care of the transition for us.

In the general case, this is for greater compatiblity with
`base` / the Haskell ecosystem at large.

My specific desire for the change comes from the fact that
`generics-eot` and some of the classes in `contravariant` go
together particularly well, and the latter works with the
`Void` from `Data.Void`.